+++ packages/webpack-config-terra/CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Unreleased
 
 * Added
-  * Added `browser` to webpack module list.
+  * Added `browser` to `resolve.mainFields` in webpack module list.
 
 ## 1.2.0 - (January 5, 2021)